Rename mutations.js to operations.js
